Eliminate unnecessary ScatterND (#2422)
Identify ScatterND(data, indices, updates) that can be replaced by
Identity(updates). This is generated by the translation of `x[:, ...] =
y` in PyTorch. The specific pattern is that the updated indices take the
form [[0], ..., [S-1]] for the first dimension, where S is the size of
the first dimension of the updated-data tensor. In effect, the
scatter-update ends up being an assignment of a new value to the entire
tensor.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>