Change loop order during rewrite (#2427)
Change the loop order when applying a collection of rewrite-rules to all
nodes in graph. The order in this PR is preferable, for a couple of
reasons.

First: rules often have mutual dependences, with one rule enabling
another. This loop-ordering handles such dependences better, and does so
more efficiently (requiring fewer iterations of yet another outer loop
to invoke the rewriter multiple times).

It also sets it up for another optimization planned for originally (but
not yet implemented): For patterns that end with a definite op (like
ScatterND), which are most of the rules/patterns, we can create a
dictionary mapping the op-identifier to rules applicable to that op.
Then, it is sufficient to iterate only over rules applicable to current
node's op.

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>