Add reproduction test case for incorrect slice rewrite and add potential fix (#2478)
This adds a reproduction of the rule introduced in
f42c2bbfd31edc99a849e0381ae4992da32479de leading to an incorrect rewrite
of the graph.

The original rule does not consider the step parameter, which can
influence the result of a `Slice` to be the identity even when input and
output shape are equivalent.

The potential fix seems to be to not apply the rule on `step != 1`,
therefore the second commit adds this to the original rule
implementation.

---------

Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>