[ort_fusuion] Support fp16 in rms_norm fusion (#2491)
In RMSNorm, there are compute_type and target_type, which we run the
computation on compute_type and then convert it back to target_type
after RMSNorm.

Typical example can be found in RMSNorm class in LLMs, like in GPT-OSS:
https://github.com/huggingface/transformers/blob/52c6c1bb6e27ca87c4faede34a4c2a7404c17c4d/src/transformers/models/gpt_oss/modeling_gpt_oss.py#L54

Therefore, we need to take op.Cast into pattern consideration.