Introduce layer-norm fusion (#2492)
Introduce layer-norm fusion rules, along with a couple of test cases.

This is just the first version. TO DO:
* We need improved infrastructure for ONNX fusions to handle opset
dependence. LayerNorm exists in ONNX since opset 17. For now the fusion
rule exists, but it is not automatically called yet (but users can
invoke it themselves).
* If users want to use opsets < 17, this could be done as an ORT fusion
using ORT contrib op LayerNorm.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>