[torchlib] Improve handling of SymInt[] (#2522)
Previously sizes coming in as `SymInt[]` are first concatenated as INT64
then used. This created inefficiencies where we could not process any
static dims from the size list and had to treat the whole shape as
dynamic. In aten_expand, this meant we needed to add `Abs` on the shape.

This change updates the functions that take `SymInt[]` such that they
are no longer turned into INT64 first. I updated aten_expand to process
constant `-1` values so an `Abs` is not required. I also added a helper
`merge_dims` to create constants for consecutive constant dims first
before concatinating.

---------

Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>