Separated implementation of aten::scatter overloads (#2605)
close #2601 #2602 

This PR refactors the implementation of `aten::scatter` overloads,
improving the clarity of the ONNX output generated by
`aten::scatter.src.`
I've also added new tests to verify the correctness of these changes. To
make the added tests pass, I needed to also address the issue reported
in #2602, which is included in this PR's diff.

Signed-off-by: Linsho Kaku <linsho@preferred.jp>