Assign names for None dims
When onnx shape inference is run on symbolic input dims, it will not handle the dim name propagation and instead create a None. As long as we rely on the current version onnx shape inference there is not better information we can get.

However, since in the optimizer we also have some custom shape propagator implemented (e.g. for Identity) that will propagate sym dims, we should encode the equivalents for those dimensions as much as possible.

This PR assigns a string to all None dims produced by onnx shape inference, so that the string names can get propagated when possible by the optimizer.

Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>