Fix scalar constant check (#2672)
Fix scalar constant check.

TODO: for some optimizations, generalizations are possible, but they
must be done on a rule-by-rule basis. Eg., for eliminating an addition
of zero to x: eliminating this is always safe zero is a scalar, but if
it is multi-dimensional, then it is safe if its rank is less than that
of x.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>