[torchlib] Consolidate all overloads and prevent new ones from being created (#2621)
This PR implements https://github.com/microsoft/onnxscript/issues/2580
by combining all overloads in torchlib and remove the ability to
register new ones. It is done in a BC compatible fashion and should work
with released versions of PyTorch.

From now on all logic for a single aten OpOverload should be implemented
by a single torchlib function to ensure 1-to-1 mapping.

---------

Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>